+++ src/test/extension_symb.test.ts
@@ -21,7 +21,7 @@ describe("Global Extension Tests", function () {
             // tslint:disable-next-line:no-unused-expression
             expect(document).to.not.be.undefined;
         });
-        it("Should find the symbol definition", async () => {
+        it("Should find a symbol definition", async () => {
             let locations = await vscode.commands.executeCommand('vscode.executeDefinitionProvider', uri, new Position(341, 33));
             // tslint:disable-next-line:no-unused-expression
             expect(locations).to.not.be.undefined;
@@ -29,6 +29,17 @@ describe("Global Extension Tests", function () {
                 let l = locations as Array<Location>;
                 expect(l.length).to.be.equal(1);
                 expect(l[0].range).to.be.eql(new Range(new Position(25, 0), new Position(25, 11)));
+                expect(l[0].uri.fsPath).to.be.eql(uri.fsPath);
+            }
+        });
+        it.only("Should find all the symbol references", async () => {
+            let locations = await vscode.commands.executeCommand('vscode.executeReferenceProvider', uri, new Position(341, 33));
+            // tslint:disable-next-line:no-unused-expression
+            expect(locations).to.not.be.undefined;
+            if (locations) {
+                let l = locations as Array<Location>;
+                expect(l.length).to.be.equal(12);
+                expect(l[0].range).to.be.eql(new Range(new Position(341, 28), new Position(341, 39)));
                 expect(l[0].uri.fsPath).to.be.eql(uri.fsPath);
             }
         });