+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### [0.8.0]
-- [Issue #19](https://github.com/prb28/vscode-amiga-assembly/issues/19) * represents the current pc in a operand symbol
+- [Issue #23](https://github.com/prb28/vscode-amiga-assembly/issues/23) Add a reference provider
+- [Issue #22](https://github.com/prb28/vscode-amiga-assembly/issues/22) Add a definition provider
+- [Issue #19](https://github.com/prb28/vscode-amiga-assembly/issues/19) * may represents the current pc in a operand symbol
 - [Feat #18](https://github.com/prb28/vscode-amiga-assembly/issues/18) Accept octal numbers 
 - [Feat #16](https://github.com/prb28/vscode-amiga-assembly/issues/16) Add VASM directives to documentation